POC -- safari extension
This works fairly cleanly for new tabs / new windows. Not sure how to
distinguish.

The popup interface is kinda corny...but could be fun if you it we can
still restrict cors access to the resources. maybe not such a big deal
to have non-branded sites request the site. Would make for an easier
embedable extension widget:

<script id="24hhpwidget" src="https://widget.24HourHomepage.com" />